llm-d-inference-scheduler-k8s/pyproject.toml
# Copyright 2026 Rob Gibbon
# See LICENSE file for licensing details.

[project]
name = "llm-d-inference-scheduler-k8s"
version = "0.0.1"
requires-python = ">=3.10"

# Dependencies of the charm code
# You should list the dependencies of the code in src/, including any charm libraries from PyPI.
# We recommend using uv to maintain this list. For example, 'uv add charmlibs-pathops'.
# If your code uses any libraries from Charmhub, don't list those libraries here. Instead, add a
# 'charm-libs' block in charmcraft.yaml, run `charmcraft fetch-libs` to download the libraries,
# then inspect the libraries for dependencies specified in PYDEPS. List those dependencies here.
dependencies = [
    "cosl>=1.6.1",
    "lightkube>=0.19.1",
    "ops>=3,<4",
    "Jinja2>=3.1.2",
    "pydantic>=2.0",
]

[dependency-groups]
# Dependencies of linting and static type checks
lint = [
    "ruff",
    "codespell",
    "pyright",
]
# Dependencies of unit tests
unit = [
    "coverage[toml]",
    "ops[testing]",
    "pytest",
]
# Dependencies of integration tests
integration = [
    "jubilant",
    "pytest",
    "PyYAML",
]
dev = [
    "pytest-cov>=7.1.0",
]

# Testing tools configuration
[tool.coverage.run]
branch = true

[tool.coverage.report]
show_missing = true

[tool.pytest.ini_options]
minversion = "6.0"
log_cli_level = "INFO"

# Linting tools configuration
[tool.ruff]
line-length = 99
lint.select = ["E", "W", "F", "C", "N", "D", "I001"]
lint.ignore = [
    "D105",
    "D107",
    "D203",
    "D204",
    "D213",
    "D215",
    "D400",
    "D404",
    "D406",
    "D407",
    "D408",
    "D409",
    "D413",
]
lint.per-file-ignores = {"tests/*" = ["D100","D101","D102","D103","D104"]}
extend-exclude = ["__pycache__", "*.egg_info"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.codespell]
skip = "build,lib,venv,icon.svg,.tox,.git,.mypy_cache,.ruff_cache,.coverage"

[tool.pyright]
include = ["src", "tests"]
